feat: add client database
