--- back-end/src/main/java/com/helesto/api/ResponseReadClient.java
@@ -1,45 +1,0 @@
-package com.helesto.api;
-
-public class ResponseReadClient {
-
-    private String name;
-
-    private String birthDate;
-
-    private String email;
-
-    private long phoneNumber;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getBirthDate() {
-        return birthDate;
-    }
-
-    public void setBirthDate(String birthDate) {
-        this.birthDate = birthDate;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public long getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(long phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-
-}
