+++ back-end/src/main/resources/application.properties
@@ -7,3 +7,31 @@ quarkus.log.level=INFO
 
 # Cross-origin resource sharing (CORS)
 quarkus.http.cors=true
+
+# Database
+quarkus.datasource.db-kind=postgresql
+quarkus.hibernate-orm.database.generation=none
+quarkus.hibernate-orm.sql-load-script=no-file
+quarkus.hibernate-orm.log.sql=true
+
+#################################################################
+### development mode ###
+#################################################################
+
+%dev.quarkus.datasource.jdbc.url=jdbc:postgresql://localhost:5432/usersdb
+%dev.quarkus.datasource.username=postgres
+%dev.quarkus.datasource.password=postgres
+%dev.quarkus.hibernate-orm.database.generation=drop-and-create
+%dev.quarkus.hibernate-orm.sql-load-script=import.sql
+%dev.quarkus.hibernate-orm.log.sql=true
+
+#################################################################
+### test mode ###
+#################################################################
+
+%test.quarkus.datasource.db-kind=h2
+%test.quarkus.datasource.jdbc.url=jdbc:h2:mem:default
+%test.quarkus.datasource.username=default
+%test.quarkus.hibernate-orm.database.generation=drop-and-create
+%test.quarkus.hibernate-orm.sql-load-script=import.sql
+%test.quarkus.hibernate-orm.log.sql=true
