+++ back-end/src/main/resources/import.sql
@@ -1,0 +1,9 @@
+INSERT INTO CLIENT (ID, NAME, BIRTH_DATE, EMAIL, PHONE_NUMBER) 
+VALUES ( nextval('CLIENT_ID_SEQUENCE'), 'John Smith','1983-03-25','john@tst.com', 5561984881234);
+
+INSERT INTO CLIENT (ID, NAME, BIRTH_DATE, EMAIL, PHONE_NUMBER) 
+VALUES ( nextval('CLIENT_ID_SEQUENCE'), 'William','1988-03-28','william@tst.com', 5592743881234);
+
+INSERT INTO CLIENT (ID, NAME, BIRTH_DATE, EMAIL, PHONE_NUMBER) 
+VALUES ( nextval('CLIENT_ID_SEQUENCE'), 'Marcus','1995-03-28','marcus@tst.com', 5592743789564);
+
